heavily refactor unity assets
preparing for adding packed bend values back
